Add files via upload
This is a semi-rough implementation of an assumed fill algorithm into the PM Randomizer. I've only tested this with race settings (and updated the default settings to race settings with Assumed Fill on); further testing is certainly necessary before this could be pushed to anything mainstream. This is significantly slower than the old forward fill, but I have at least implemented a bit of a cheat with Star Pieces/Letters to cut down a bit (current gen times are averaging a bit more than 6x the forward fill gen times, improved from the 10x of a more naïve implementation). I have also not actually tested in-game beyond seeing the ROM loads and the first shop has the right items in it.