fix: trap and non-trap items treated the same item
Due to a missing __eq__ function in the Item class, the logic module
considered a trapped version and a not trapped version of an item to be
the same instance of an item during placing of partner upgrades.
This caused weird behavior with the 'remove' function of collections,
and sometimes turned trapped partner upgrades into not trapped ones,
causing duplicate upgrades in the item pool.